Move http history related classes into a HttpHistory namespace
